fix : refactor game states
복잡한 관계를 가진 game state 구조 정리
