+++ src/main/java/domain/state/BustGameState.java
@@ -2,7 +2,7 @@ package domain.state;
 
 import domain.Hand;
 
-public class BustGameState extends EndGameState {
+public class BustGameState extends FinishedGameState {
     public BustGameState(Hand hand) {
         super(hand);
     }
