+++ src/main/java/domain/state/GameState.java
@@ -1,24 +1,16 @@
 package domain.state;
 
 import domain.Card;
-import domain.GameResult;
 import domain.Hand;
 import java.util.List;
 import java.util.function.Supplier;
 
 public interface GameState {
-    static GameState createPlayerInitialGameState(Hand hand) {
+    static GameState createInitialGameState(Hand hand) {
         if (hand.isBlackJack()) {
             return new BlackJackGameState(hand);
         }
-        return new PlayerPlayableGameState(hand);
-    }
-
-    static GameState createDealerInitialGameState(Hand hand) {
-        if (hand.isBlackJack()) {
-            return new BlackJackGameState(hand);
-        }
-        return new DealerPlayableGameState(hand);
+        return new RunningGameState(hand);
     }
 
     GameState hit(Supplier<Card> cardSupplier);
@@ -36,6 +28,4 @@ public interface GameState {
     List<Card> showOwnCards();
 
     int getCardsSum();
-
-    GameResult compare(GameState playerState, GameState dealerState);
 }
