+++ src/main/java/domain/state/StayGameState.java
@@ -2,7 +2,7 @@ package domain.state;
 
 import domain.Hand;
 
-public class StayGameState extends EndGameState {
+public class StayGameState extends FinishedGameState {
     public StayGameState(Hand hand) {
         super(hand);
     }
